requirements.txt
flask
requests
opencv-python
ultralytics
python3-pika